+++ cypress/cypress/component/PanelSidebar/PanelSidebar.cy.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable max-lines */
 import { PropsWithChildren } from "react";
 import {
   PanelSideBarProvider,
@@ -9,7 +8,7 @@ import {
   PanelItemOnSideBarCollapseOptions,
 } from "react-pattern-ui";
 import { faBars, faCogs, faInfo, faHome, faPerson } from "@fortawesome/free-solid-svg-icons";
-import { mount } from "cypress/react18";
+import { mount } from "cypress/react";
 
 type AppRoutes = "home" | "settings" | "dropdownTest" | "dropdown-test1" | "dropdown-test2" | "info";
 type TSideBarMenuItem = PanelItem<AppRoutes>;