+++ src/lib/Layout/PanelSideBarLayout/PanelSideBarNavbar.tsx
@@ -1,7 +1,7 @@
 import { faBars } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { ReactNode } from "react";
-import { Nav, NavItem } from "reactstrap";
+import { Nav, NavItem } from "@neolution-ch/reactstrap";
 import classNames from "classnames";
 import { SidebarProps } from "./PanelSideBar/Context/PanelSideBarContextProps";